feat: make tensorflow and pytorch optional
