+++ vehicle-python/pyproject.toml
@@ -17,23 +17,30 @@ license-files = [ "LICENSE" ]
 dynamic = ["version"]
 requires-python = ">=3.10,<3.14"
 dependencies = [
-    "typing-extensions>=4.6,<5",
-    "jaxtyping>=0.2, <0.3",
-    "pytest>=8.4.2",
-    "pre-commit>=4.3.0",
+  "typing-extensions>=4.6,<5",
+  "jaxtyping>=0.2, <0.3",
 ]
 
 [project.optional-dependencies]
-test = ["pytest >=7.1,<9", "packaging >=23", "pygments >=2.14, <3"]
-mypy = ["types_setuptools >=45", "types_pygments >=2.14,<3"]
-pygments = ["pygments >=2.14,<3"]
-numpy = ["numpy >=1.21,<3"]
-jax = ["jax >=0.4.26"]
+test = [
+  "pytest>=7.1,<9",
+  "packaging>=23",
+  "pygments>=2.14,<3",
+  "numpy>=1.21,<3",
+]
+mypy = [
+  "types-setuptools>=45",
+  "types-pygments>=2.14,<3",
+]
+pygments = ["pygments>=2.14,<3"]
+numpy = ["numpy>=1.21,<3"]
+jax = ["jax>=0.4.26"]
+dev = ["pre-commit>=4.3.0"]
 tensorflow = [
-  "tensorflow >=2,<3; (python_version>='3.10' and python_version<'3.13') and ((sys_platform=='linux' and (platform_machine=='x86_64' or platform_machine=='aarch64')) or (sys_platform=='darwin' and platform_machine=='x86_64') or (sys_platform=='win32' and platform_machine=='AMD64'))"
+  "tensorflow>=2,<3; (python_version >= '3.10' and python_version < '3.13') and ((sys_platform == 'linux' and (platform_machine == 'x86_64' or platform_machine == 'aarch64')) or (sys_platform == 'darwin' and platform_machine == 'x86_64') or (sys_platform == 'win32' and platform_machine == 'AMD64'))",
 ]
 pytorch = [
-  "torch >=2.0,<3.0; python_version>='3.10' and python_version<'3.13'"
+  "torch>=2,<3; python_version >= '3.10' and python_version < '3.13'",
 ]
 
 [project.scripts]
@@ -100,6 +107,31 @@ testpaths = ["tests"]
 
 [tool.setuptools.dynamic]
 version = { attr = "vehicle_lang._version.VERSION" }
+
+[tool.uv]
+default-groups = []
+
+[dependency-groups]
+test = [
+  "pytest>=7.1,<9",
+  "packaging>=23",
+  "pygments>=2.14,<3",
+  "numpy>=1.21,<3",
+]
+mypy = [
+  "types-setuptools>=45",
+  "types-pygments>=2.14,<3",
+]
+pygments = ["pygments>=2.14,<3"]
+numpy = ["numpy>=1.21,<3"]
+jax = ["jax>=0.4.26"]
+dev = ["pre-commit>=4.3.0"]
+tensorflow = [
+  "tensorflow>=2,<3; (python_version >= '3.10' and python_version < '3.13') and ((sys_platform == 'linux' and (platform_machine == 'x86_64' or platform_machine == 'aarch64')) or (sys_platform == 'darwin' and platform_machine == 'x86_64') or (sys_platform == 'win32' and platform_machine == 'AMD64'))",
+]
+pytorch = [
+  "torch>=2,<3; python_version >= '3.10' and python_version < '3.13'",
+]
 
 [tool.tox]
 legacy_tox_ini = """
