+++ vehicle-python/src/vehicle_lang/__init__.py
@@ -1,21 +1,13 @@
 from typing import List
 
-from . import compile, session
+from . import loss, session
 from ._version import VERSION
 from .check import check
-from .compile import (
-    DefaultPyTorchSampler,
-    DefaultTensorFlowSampler,
-    PyTorchSampler,
-    TensorFlowSampler,
-    call_vehicle,
-    compile_specification,
-    load_specification,
-)
-from .compile.error import VehicleBuiltinUnsupported, VehiclePropertyNotFound
+from .compile import call_vehicle, compile_specification
 from .error import VehicleError, VehicleInternalError
 from .export import export_to_solver
 from .list import list
+from .loss.error import VehicleBuiltinUnsupported, VehiclePropertyNotFound
 from .session.error import VehicleSessionClosed, VehicleSessionUsed
 from .typing import (
     DeclarationName,
@@ -36,12 +28,8 @@ __all__: List[str] = [
     "check",
     # Compile
     "compile_specification",
-    "load_specification",
-    # Samplers
-    "PyTorchSampler",
-    "DefaultPyTorchSampler",
-    "TensorFlowSampler",
-    "DefaultTensorFlowSampler",
+    # Loss helpers
+    "loss",
     # Call Vehicle
     "call_vehicle",
     # Verify
@@ -58,8 +46,8 @@ __all__: List[str] = [
     "VehicleError",
     "VehicleSessionClosed",
     "VehicleSessionUsed",
-    "VehicleBuiltinUnsupported",
     "VehicleInternalError",
+    "VehicleBuiltinUnsupported",
     "VehiclePropertyNotFound",
     # Argument types
     "DeclarationName",
