+++ vehicle-python/src/vehicle_lang/compile/__init__.py
@@ -1,13 +1,9 @@
-from collections import defaultdict
 from pathlib import Path
-from typing import Any, Iterable, Mapping, Optional
+from typing import Any, Iterable, Optional
 
 from .. import session as session
-from ..error import VehicleError as VehicleError
-from ..typing import ITP, DeclarationName, DifferentiableLogic, LossBackend, QueryFormat
-from . import _ast, error
-from .pytorch import DefaultPyTorchSampler, PyTorchSampler
-from .tensorflow import DefaultTensorFlowSampler, TensorFlowSampler
+from ..error import VehicleError
+from ..typing import ITP, DeclarationName, DifferentiableLogic, QueryFormat
 
 
 def compile_specification(
@@ -79,70 +75,6 @@ def compile_specification(
     return out
 
 
-def load_specification(
-    path: str | Path,
-    *,
-    backend: LossBackend,
-    logic: DifferentiableLogic = DifferentiableLogic.DL2,
-    samplers: Mapping[str, Any] | None = None,
-    declarations: Iterable[DeclarationName] = (),
-    declaration_context: dict[str, Any] | None = None,
-) -> dict[str, Any]:
-    """
-    Load a loss function from a Vehicle specification.
-
-    If samplers is not provided, a default sampler will be used for each
-    quantified variable encountered in the specification.
-    """
-    if declaration_context is None:
-        declaration_context = {}
-
-    # Create a defaultdict with the appropriate default sampler based on backend
-    if samplers is None:
-        match backend:
-            case LossBackend.TensorFlow:
-                from .tensorflow.samplers import DefaultTensorFlowSampler
-
-                default_sampler_tf = DefaultTensorFlowSampler()
-                samplers = defaultdict(lambda: default_sampler_tf.get_loss)
-            case LossBackend.PyTorch:
-                from .pytorch.samplers import DefaultPyTorchSampler
-
-                default_sampler_pt = DefaultPyTorchSampler()
-                samplers = defaultdict(lambda: default_sampler_pt.get_loss)
-            case _:
-                raise NotImplementedError(f"Backend {backend} not supported.")
-
-    program = _ast.load(
-        path,
-        target=logic,
-        declarations=declarations,
-    )
-
-    match backend:
-        case LossBackend.TensorFlow:
-            from .tensorflow._translation import TensorFlowTranslation
-
-            result = TensorFlowTranslation().compile(
-                program=program,
-                path=path,
-                declaration_context=declaration_context,
-                samplers=samplers,
-            )
-        case LossBackend.PyTorch:
-            from .pytorch._translation import PyTorchTranslation
-
-            result = PyTorchTranslation().compile(
-                program=program,
-                path=path,
-                declaration_context=declaration_context,
-                samplers=samplers,
-            )
-        case _:
-            raise NotImplementedError(f"Backend {backend} not supported.")
-    return result
-
-
 def call_vehicle(args: list[str]) -> str:
     # Call Vehicle
     exec, out, err, _ = session.check_output(args)
@@ -157,11 +89,5 @@ def call_vehicle(args: list[str]) -> str:
 
 __all__ = [
     "compile_specification",
-    "load_specification",
     "call_vehicle",
-    "PyTorchSampler",
-    "DefaultPyTorchSampler",
-    "TensorFlowSampler",
-    "DefaultTensorFlowSampler",
-    "error",
 ]
