+++ vehicle-python/src/vehicle_lang/loss/_pytorch/_translation.py
@@ -1,27 +1,27 @@
 import ast as py
 from dataclasses import dataclass
 
-from ..python import PythonTranslation
-from ._builtins import TensorFlowBuiltins
+from .._python import PythonTranslation
+from ._builtins import PyTorchBuiltins
 
 # Create proper Python AST provenance (different from Vehicle provenance)
 PY_MISSING = {"lineno": 0, "col_offset": 0}
 
 ################################################################################
-### TensorFlow Translation
+### PyTorch Translation
 ################################################################################
 
 
 @dataclass(frozen=True, init=False)
-class TensorFlowTranslation(PythonTranslation):
+class PyTorchTranslation(PythonTranslation):
     def __init__(self) -> None:
         super().__init__(
-            builtins=TensorFlowBuiltins(),
+            builtins=PyTorchBuiltins(),
             module_header=[
                 py.Import(
                     names=[
                         py.alias(
-                            name="tensorflow",
+                            name="torch",
                             asname=None,
                             lineno=0,
                             col_offset=0,
