+++ vehicle-python/src/vehicle_lang/pygments/_external/__init__.py
@@ -1,4 +1,4 @@
-# File generated by the BNF Converter (bnfc 2.9.4.1).
+# File generated by the BNF Converter (bnfc 2.9.5).
 
 import pygments.lexer
 from pygments.token import *
@@ -9,7 +9,7 @@ __all__ = ["ExternalLexer"]
 class ExternalLexer(pygments.lexer.RegexLexer):
     name = "External"
     aliases = ["external"]
-    KEYWORDS = ["in", "type"]
+    KEYWORDS = ["in", "record", "type", "where"]
 
     def get_tokens_unprocessed(self, text):
         for index, token, value in super(ExternalLexer, self).get_tokens_unprocessed(
@@ -27,12 +27,10 @@ class ExternalLexer(pygments.lexer.RegexLexer):
             (r"True|False", Name),
             (r"(\d)+", Name),
             (r"(\d)+\.(\d)+", Name),
-            (r"@network", Name),
-            (r"@dataset", Name),
-            (r"@parameter", Name),
-            (r"@property", Name),
+            (r"@(network|dataset|parameter|property|tensor|instance)", Name),
             (r"@postulate", Name),
-            (r"@noinline", Name),
+            (r"\{", Name),
+            (r"\}", Name),
             (r"->", Name),
             (r"forallT", Name),
             (r"if", Name),
@@ -45,12 +43,12 @@ class ExternalLexer(pygments.lexer.RegexLexer):
             (r"Type", Name),
             (r"Unit", Name),
             (r"Bool", Name),
-            (r"Nat", Name),
-            (r"Int", Name),
-            (r"Rat", Name),
-            (r"Vector", Name),
-            (r"List", Name),
             (r"Index", Name),
+            (r"Nat", Name),
+            (r"Real", Name),
+            (r"List", Name),
+            (r"Vector", Name),
+            (r"Tensor", Name),
             (r"forall", Name),
             (r"exists", Name),
             (r"foreach", Name),
@@ -64,22 +62,33 @@ class ExternalLexer(pygments.lexer.RegexLexer):
             (r"<", Name),
             (r">=", Name),
             (r">", Name),
+            (r"==\.", Name),
+            (r"!=\.", Name),
+            (r"<=\.", Name),
+            (r"<\.", Name),
+            (r">=\.", Name),
+            (r">\.", Name),
             (r"\*", Name),
             (r"/", Name),
             (r"\+", Name),
             (r"-", Name),
+            (r"min", Name),
+            (r"max", Name),
             (r"nil", Name),
             (r"::", Name),
             (r"\[", Name),
             (r"\]", Name),
-            (r"::v", Name),
             (r"!", Name),
             (r"map", Name),
             (r"fold", Name),
-            (r"dfold", Name),
-            (r"indices", Name),
-            (r"fromNat", Name),
-            (r"fromInt", Name),
+            (r"const", Name),
+            (r"reduceOr", Name),
+            (r"reduceAnd", Name),
+            (r"reduceAdd", Name),
+            (r"reduceMul", Name),
+            (r"reduceMin", Name),
+            (r"reduceMax", Name),
+            (r"HasLeq", Name),
             (r"HasEq", Name),
             (r"HasNotEq", Name),
             (r"HasAdd", Name),
@@ -87,11 +96,14 @@ class ExternalLexer(pygments.lexer.RegexLexer):
             (r"HasMul", Name),
             (r"HasFold", Name),
             (r"HasMap", Name),
+            (r"IsTensorType", Name),
+            (r"name", Name),
+            (r"infer", Name),
             (r"[a-zA-Z](_|\d|[a-zA-Z])*", Name),
             (r"\?(_|\d|[a-zA-Z])*", Name),
-            (r"(\d|[a-zA-Z])+", Name),
+            (r"\.[a-zA-Z](_|\d|[a-zA-Z])*", Name),
             (r"[a-zA-Z]([a-zA-Z]|\d|_|\')*", Name),
-            (r"\(|\)|\{|\}|\{\{|\}\}|=|,|\(\)|;", Operator),
+            (r"@0|\(|\)|\{|\}|\{\{|\}\}|=|,|\(\)|;", Operator),
             (r"(\d)+", Number.Integer),
             (r"(\d)+\.(\d)+(e(-)?(\d)+)?", Number.Float),
             (r'"((.)(?<!["\\])|\\["\\nt])*"', String.Double),
