+++ vehicle-python/tests/data/test_bounded.vcl
@@ -1,6 +1,6 @@
 @network
-network : Real -> Real
+network : Tensor Real [1] -> Tensor Real [1]
 
 @property
 bounded : Bool
-bounded = forall (x : Real) . 0 < x < 1  => 0 < network x  < 1
+bounded = forall (x : Real) . 0 < x  < 1  => 0 < (network [x]) ! 0 < 1
