+++ vehicle-python/tests/data/test_bounded.vcl.tokens.golden
@@ -4,11 +4,23 @@ Token.Name	'network'
 Token.Space	' '
 Token.Punctuation	':'
 Token.Space	' '
+Token.Name	'Tensor'
+Token.Space	' '
 Token.Name	'Real'
+Token.Space	' '
+Token.Operator	'['
+Token.Literal.Number.Integer	'1'
+Token.Operator	']'
 Token.Space	' '
 Token.Operator	'->'
 Token.Space	' '
+Token.Name	'Tensor'
+Token.Space	' '
 Token.Name	'Real'
+Token.Space	' '
+Token.Operator	'['
+Token.Literal.Number.Integer	'1'
+Token.Operator	']'
 Token.Space	'\n\n'
 Token.Keyword.Declaration	'@property'
 Token.Space	'\n'
@@ -39,7 +51,7 @@ Token.Space	' '
 Token.Operator	'<'
 Token.Space	' '
 Token.Name	'x'
-Token.Space	' '
+Token.Space	'  '
 Token.Operator	'<'
 Token.Space	' '
 Token.Literal.Number.Integer	'1'
@@ -50,10 +62,18 @@ Token.Literal.Number.Integer	'0'
 Token.Space	' '
 Token.Operator	'<'
 Token.Space	' '
+Token.Punctuation	'('
 Token.Name	'network'
 Token.Space	' '
+Token.Operator	'['
 Token.Name	'x'
-Token.Space	'  '
+Token.Operator	']'
+Token.Punctuation	')'
+Token.Space	' '
+Token.Operator	'!'
+Token.Space	' '
+Token.Literal.Number.Integer	'0'
+Token.Space	' '
 Token.Operator	'<'
 Token.Space	' '
 Token.Literal.Number.Integer	'1'
