+++ vehicle-python/tests/test_lossdl2_exec_tf_mnist_robustness.py
@@ -1,10 +1,11 @@
 # 2024-04-16: Disable end-to-end tests
 if False:
     from pathlib import Path
-    from typing import Any, Callable, Dict, Tuple, cast
+    from typing import Any, Callable, cast
+
+    from typing_extensions import TypeAlias
 
     import vehicle_lang as vcl
-    from typing_extensions import TypeAlias
 
     GOLDEN_PATH = (
         Path(__file__).parent.parent
