+++ vehicle-python/tests/test_lossdl2_load.py
@@ -1,7 +1,8 @@
 from pathlib import Path
 
 import pytest
-import vehicle_lang.compile._ast as vcl_ast
+
+import vehicle_lang.loss._ast as vcl_ast
 import vehicle_lang.typing as vcl_typing
 
 GOLDEN_PATH = Path(__file__).parent / "data"
