+++ vehicle-python/tests/test_pytorch_backend.py
@@ -1,12 +1,15 @@
 """Tests for PyTorch backend compilation."""
 
 import pytest
-import torch
+
+torch = pytest.importorskip(
+    "torch", reason="PyTorch extra is required for backend tests"
+)
 
 
 def test_pytorch_builtins_basic_operations() -> None:
     """Test that PyTorchBuiltins implements basic tensor operations correctly."""
-    from vehicle_lang.compile.pytorch._builtins import PyTorchBuiltins
+    from vehicle_lang.loss._pytorch._builtins import PyTorchBuiltins
 
     builtins = PyTorchBuiltins()
 
@@ -24,7 +27,7 @@ def test_pytorch_builtins_basic_operations() -> None:
 
 def test_pytorch_builtins_advanced_operations() -> None:
     """Test advanced PyTorch builtin operations."""
-    from vehicle_lang.compile.pytorch._builtins import PyTorchBuiltins
+    from vehicle_lang.loss._pytorch._builtins import PyTorchBuiltins
 
     builtins = PyTorchBuiltins()
 
@@ -47,7 +50,7 @@ def test_pytorch_builtins_advanced_operations() -> None:
 
 def test_pytorch_translation_basic() -> None:
     """Test that PyTorchTranslation can be instantiated."""
-    from vehicle_lang.compile.pytorch._translation import PyTorchTranslation
+    from vehicle_lang.loss._pytorch._translation import PyTorchTranslation
 
     translation = PyTorchTranslation()
     assert translation is not None
@@ -58,7 +61,7 @@ def test_pytorch_translation_basic() -> None:
 @pytest.mark.skipif(not torch.cuda.is_available(), reason="CUDA not available")  # type: ignore[misc]
 def test_pytorch_cuda_compatibility() -> None:
     """Test that PyTorch backend works with CUDA if available."""
-    from vehicle_lang.compile.pytorch._builtins import PyTorchBuiltins
+    from vehicle_lang.loss._pytorch._builtins import PyTorchBuiltins
 
     builtins = PyTorchBuiltins()
